requirements.txt
flask
groq
opencv-python
numpy
Pillow
python-dotenv
werkzeug